rename vue component
