+++ example/src/components/arrayease.vue
